Add cleanUp method
After initializing the Connect module, the instance dispatches an event
listener for the tunnel between the host (your app) and the guest
(connect app). When instantiating a new instance of Connect, a new event
listener is created, and you will receive duplicity of events, causing
unexpected behavior or even performance issues or memory leak.

This commit creates a method to clean up the event listener of an
instance of Connect.